Working version of C# Windows Form Client interacting with a REST WebAPI service.
The following project consist of

GR.Business - Which contains business specific code. For this example. This will handle all the sorting features of generic people data.

GR.Cleint - Which contains the visual windows for user interface that the user will be able to send data over to a REST service and communicate back and forth.

GR.Contracts - Which contains data contracts that can be shared across client side and server side. This is to keep a consistent way of handling data POCOS.

GR.Data - Which contains our way of interacting with data. This current example has two ways. 1. Using an XML file as a persistence storage file. 2. Using an MDF database file that allows using SQL. This follows an interface standard that various data strageties could be implemented.

GR.Data.Tests - Which contains unit test to will test out our various data accessing strageties.

GR.Manager - Which contains classes that are responsible for things like Security and Logging. The managers are called before anything in the business will be called.

GR.Manager.Tests - Which contains a way of testing out security and the business logic.

GR.REST - Which contains the WebAPI REST controller (Action) verbs of POST, PUT, GET, DELETE